一次全面的业务细处完善与改进，包括新的用户信息未完成异常、信誉分扣减业务完成、pom文件删除重复部分、部分可能报空指针异常的if处理等。

+++ src/main/java/com/example/loan/dao/entity/RepayOverdueRecord.java
@@ -49,5 +49,5 @@ public class RepayOverdueRecord {
 
     //本次扣减信用分
     @Column(name = "credit_decrease")
-    private float creditDecrease;
+    private Integer creditDecrease;
 }
